requirements.txt
Cerberus>=1.1
click>=6.2
Jinja2>=2.8
PyYAML>=3.1
docker-py>=1.6.0
requests>2.11
requests-toolbelt
mock>=2.0.0
pexpect>=4.2.1